Hide UIKit in a compiler branch
